Add working functionality and restructure code
